deer-hunting-game/requirements.txt
requests==2.31.0
python-engineio==3.14.2
python-socketio==4.6.1